Update index.md
Please verify that I have edited away from the core intent and capability of Buoy. I would also recommend using a different screenshot because it's still distracting to me even after engaging the copy for hours. A better screenshot might include a map and narrative of folks using it for natural disaster response or street medics during a protest event if you want to make the usefulness to the movement more clear. The current one begs too many questions (doesn't explicit state that the alert is being place for copWATCHING and not anything else…I'm assuming that this is what's happening). I apologize in advance. Editing sucks to do but I think it's good to have my non-developer take on it. YOU DO NOT HAVE TO USE A SINGLE RECOMMENDATION MY FEELINGS WILL NOT BE HURT I PROMISE. I do a lot of reading and study on copy, branding, and marketing. I'm not a professional but I'm passionate about delivery especially for important tools like this one. At any rate, I hope this is helpful. Please reach out if you have questions (or not if you never wanna hear from me again lol). I appreciate what you're doing!